requirements.txt
requests
beautifulsoup4
feedparser
jinja2
scrapling
python-telegram-bot
schedule
piper-tts
